Participants in a MOOC discussion forum

The participant table accompanying mooc_posts: every person who appears as a sender or a receiver there, with the self-reported experience level the chapter recodes into Expert, Student and Teacher and then uses as the mixing attribute.

Usage

mooc_people

Format

A data frame with 445 rows and 2 columns:

name

Character. Participant identifier, matching the sender and receiver columns of mooc_posts.

experience

Integer. Self-reported experience level: 1 expert, 2 student, 3 teacher.

Source

As mooc_posts.

See also

Examples

participants <- dynet(mooc_posts, from = "sender", to = "receiver",
                      time = "timestamp", thread = "discussion",
                      nodes = mooc_people)
#> Dropped 86 self-loop event(s). Use loops = TRUE to keep them.
as.data.frame(participants, what = "nodes")
#>     name experience
#> 1      1          1
#> 2      2          1
#> 3      3          2
#> 4      4          2
#> 5      5          3
#> 6      6          1
#> 7      7          2
#> 8      8          1
#> 9      9          1
#> 10    10          2
#> 11    11          3
#> 12    12          3
#> 13    13          2
#> 14    14          1
#> 15    15          3
#> 16    16          1
#> 17    17          1
#> 18    18          1
#> 19    19          3
#> 20    20          1
#> 21    21          1
#> 22    22          1
#> 23    23          3
#> 24    24          2
#> 25    25          2
#> 26    26          3
#> 27    27          1
#> 28    28          3
#> 29    29          2
#> 30    30          3
#> 31    31          1
#> 32    32          1
#> 33    33          3
#> 34    34          3
#> 35    35          2
#> 36    36          2
#> 37    37          2
#> 38    38          1
#> 39    39          2
#> 40    40          3
#> 41    41          1
#> 42    42          3
#> 43    43          3
#> 44    44          2
#> 45    45          2
#> 46    46          2
#> 47    47          3
#> 48    48          3
#> 49    49          3
#> 50    50          3
#> 51    51          3
#> 52    52          3
#> 53    53          3
#> 54    54          2
#> 55    55          1
#> 56    56          3
#> 57    57          3
#> 58    58          2
#> 59    59          1
#> 60    60          2
#> 61    61          3
#> 62    62          1
#> 63    63          1
#> 64    64          2
#> 65    65          2
#> 66    66          2
#> 67    67          3
#> 68    68          3
#> 69    69          3
#> 70    70          1
#> 71    71          3
#> 72    72          2
#> 73    73          3
#> 74    74          3
#> 75    75          3
#> 76    76          3
#> 77    77          2
#> 78    78          3
#> 79    79          1
#> 80    80          2
#> 81    81          3
#> 82    82          1
#> 83    83          3
#> 84    84          1
#> 85    85          1
#> 86    86          3
#> 87    87          2
#> 88    88          1
#> 89    89          3
#> 90    90          3
#> 91    91          2
#> 92    92          3
#> 93    93          1
#> 94    94          1
#> 95    95          1
#> 96    96          2
#> 97    97          1
#> 98    98          2
#> 99    99          3
#> 100  100          2
#> 101  101          1
#> 102  102          3
#> 103  103          2
#> 104  104          3
#> 105  105          2
#> 106  106          3
#> 107  107          2
#> 108  108          1
#> 109  109          3
#> 110  110          1
#> 111  111          2
#> 112  112          2
#> 113  113          2
#> 114  114          2
#> 115  115          2
#> 116  116          3
#> 117  117          1
#> 118  118          1
#> 119  119          2
#> 120  120          3
#> 121  121          2
#> 122  122          3
#> 123  123          2
#> 124  124          2
#> 125  125          3
#> 126  126          2
#> 127  127          2
#> 128  128          1
#> 129  129          2
#> 130  130          3
#> 131  131          2
#> 132  132          1
#> 133  133          3
#> 134  134          1
#> 135  135          1
#> 136  136          3
#> 137  137          3
#> 138  138          2
#> 139  139          2
#> 140  140          3
#> 141  141          3
#> 142  142          2
#> 143  143          2
#> 144  144          2
#> 145  145          1
#> 146  146          1
#> 147  147          2
#> 148  148          2
#> 149  149          1
#> 150  150          1
#> 151  151          2
#> 152  152          2
#> 153  153          1
#> 154  154          3
#> 155  155          3
#> 156  156          3
#> 157  157          1
#> 158  158          3
#> 159  159          2
#> 160  160          3
#> 161  161          3
#> 162  162          3
#> 163  163          1
#> 164  164          2
#> 165  165          3
#> 166  166          1
#> 167  167          3
#> 168  168          3
#> 169  169          3
#> 170  170          1
#> 171  171          2
#> 172  172          2
#> 173  173          2
#> 174  174          2
#> 175  175          3
#> 176  176          2
#> 177  177          2
#> 178  178          2
#> 179  179          2
#> 180  180          3
#> 181  181          1
#> 182  182          1
#> 183  183          2
#> 184  184          1
#> 185  185          2
#> 186  186          1
#> 187  187          2
#> 188  188          3
#> 189  189          3
#> 190  190          1
#> 191  191          3
#> 192  192          3
#> 193  193          2
#> 194  194          3
#> 195  195          3
#> 196  196          1
#> 197  197          3
#> 198  198          2
#> 199  199          3
#> 200  200          2
#> 201  201          1
#> 202  202          2
#> 203  203          3
#> 204  204          2
#> 205  205          1
#> 206  206          2
#> 207  207          1
#> 208  208          2
#> 209  209          3
#> 210  210          2
#> 211  211          2
#> 212  212          3
#> 213  213          2
#> 214  214          1
#> 215  215          1
#> 216  216          3
#> 217  217          3
#> 218  218          1
#> 219  219          3
#> 220  220          3
#> 221  221          1
#> 222  222          3
#> 223  223          2
#> 224  224          3
#> 225  225          3
#> 226  226          3
#> 227  227          2
#> 228  228          1
#> 229  229          2
#> 230  230          3
#> 231  231          3
#> 232  232          3
#> 233  233          2
#> 234  234          3
#> 235  235          2
#> 236  236          1
#> 237  237          2
#> 238  238          1
#> 239  239          1
#> 240  240          1
#> 241  241          1
#> 242  242          1
#> 243  243          1
#> 244  244          3
#> 245  245          1
#> 246  246          3
#> 247  247          2
#> 248  248          1
#> 249  249          2
#> 250  250          2
#> 251  251          2
#> 252  252          2
#> 253  253          3
#> 254  254          3
#> 255  255          2
#> 256  256          3
#> 257  257          3
#> 258  258          3
#> 259  259          3
#> 260  260          3
#> 261  261          2
#> 262  262          3
#> 263  263          2
#> 264  264          2
#> 265  265          3
#> 266  266          1
#> 267  267          3
#> 268  268          1
#> 269  269          3
#> 270  270          1
#> 271  271          2
#> 272  272          3
#> 273  273          3
#> 274  274          3
#> 275  275          1
#> 276  276          3
#> 277  277          3
#> 278  278          2
#> 279  279          3
#> 280  280          3
#> 281  281          3
#> 282  282          1
#> 283  283          2
#> 284  284          1
#> 285  285          1
#> 286  286          3
#> 287  287          1
#> 288  288          1
#> 289  289          2
#> 290  290          3
#> 291  291          1
#> 292  292          1
#> 293  293          3
#> 294  294          3
#> 295  295          1
#> 296  296          3
#> 297  297          3
#> 298  298          3
#> 299  299          3
#> 300  300          2
#> 301  301          2
#> 302  302          3
#> 303  303          3
#> 304  304          1
#> 305  305          3
#> 306  306          1
#> 307  307          2
#> 308  308          1
#> 309  309          1
#> 310  310          2
#> 311  311          2
#> 312  312          2
#> 313  313          2
#> 314  314          2
#> 315  315          3
#> 316  316          3
#> 317  317          3
#> 318  318          2
#> 319  319          2
#> 320  320          3
#> 321  321          2
#> 322  322          1
#> 323  323          1
#> 324  324          3
#> 325  325          2
#> 326  326          3
#> 327  327          1
#> 328  328          3
#> 329  329          2
#> 330  330          3
#> 331  331          2
#> 332  332          1
#> 333  333          1
#> 334  334          3
#> 335  335          3
#> 336  336          2
#> 337  337          3
#> 338  338          3
#> 339  339          3
#> 340  340          3
#> 341  341          3
#> 342  342          3
#> 343  343          1
#> 344  344          3
#> 345  345          3
#> 346  346          1
#> 347  347          2
#> 348  348          3
#> 349  349          1
#> 350  350          2
#> 351  351          2
#> 352  352          1
#> 353  353          1
#> 354  354          2
#> 355  355          2
#> 356  356          2
#> 357  357          3
#> 358  358          3
#> 359  359          3
#> 360  360          1
#> 361  361          3
#> 362  362          2
#> 363  363          3
#> 364  364          3
#> 365  365          2
#> 366  366          2
#> 367  367          1
#> 368  368          2
#> 369  369          3
#> 370  370          3
#> 371  371          2
#> 372  372          3
#> 373  373          1
#> 374  374          2
#> 375  375          3
#> 376  376          2
#> 377  377          2
#> 378  378          1
#> 379  379          2
#> 380  380          3
#> 381  381          2
#> 382  382          2
#> 383  383          3
#> 384  384          2
#> 385  385          3
#> 386  386          2
#> 387  387          3
#> 388  388          2
#> 389  389          3
#> 390  390          1
#> 391  391          3
#> 392  392          3
#> 393  393          2
#> 394  394          1
#> 395  395          2
#> 396  396          2
#> 397  397          1
#> 398  398          1
#> 399  399          2
#> 400  400          3
#> 401  401          2
#> 402  402          2
#> 403  403          3
#> 404  404          1
#> 405  405          1
#> 406  406          2
#> 407  407          3
#> 408  408          3
#> 409  409          1
#> 410  410          1
#> 411  411          3
#> 412  412          1
#> 413  413          3
#> 414  414          2
#> 415  415          3
#> 416  416          2
#> 417  417          2
#> 418  418          2
#> 419  419          3
#> 420  420          2
#> 421  421          2
#> 422  422          2
#> 423  423          3
#> 424  424          2
#> 425  425          3
#> 426  426          1
#> 427  427          3
#> 428  428          3
#> 429  429          1
#> 430  430          2
#> 431  431          2
#> 432  432          1
#> 433  433          3
#> 434  434          1
#> 435  435          1
#> 436  436          2
#> 437  437          1
#> 438  438          1
#> 439  439          1
#> 440  440          2
#> 441  444          3
#> 442  445          3